feat: replace glam with macaw, inline relevant types
This required making changes to the inlined macaw code to remove dependencies on other macaw code, which hopefully shouldn't bite us too much. We may want to optionally support `bevy_math` in future?

This also uses a Git version of `speedy` to support glam 0.27. We can hopefully switch this over to a new `speedy` release in future.